test: improve Chrome Extension coverage with CDP/MV3
