+++ tests/assets/extension-mv2-simple/content-script.js
