+++ tests/assets/extension-mv2-simple/index.js
