+++ tests/assets/extension-mv2-simple/manifest.json
