+++ tests/assets/extension-mv3-simple/content-script.js
@@ -1,0 +1,3 @@
+console.log('hey from the content-script');
+self.thisIsTheContentScript = true;
+
