+++ tests/assets/extension-mv3-simple/index.js
@@ -1,0 +1,2 @@
+// Mock script for background extension
+globalThis.MAGIC = 42;
