+++ tests/assets/extension-mv3-with-logging/background.js
