+++ tests/assets/extension-mv3-with-logging/content.js
