+++ tests/assets/extension-mv3-with-logging/manifest.json
