+++ tests/library/chromium/launcher.spec.ts
@@ -52,121 +52,6 @@ it('should open devtools when "devtools: true" option is given', async ({ browse
   await browser.close();
 });
 
-it('should return background pages', async ({ browserType, asset, isHeadlessShell }) => {
-  it.skip(isHeadlessShell, 'Headless Shell has no support for extensions');
-
-  const extensionPath = asset('simple-extension');
-  const extensionOptions = {
-    args: [
-      `--disable-extensions-except=${extensionPath}`,
-      `--load-extension=${extensionPath}`,
-    ],
-  };
-  const context = await browserType.launchPersistentContext('', extensionOptions);
-  const backgroundPages = context.backgroundPages();
-  const backgroundPage = backgroundPages.length
-    ? backgroundPages[0]
-    : await context.waitForEvent('backgroundpage');
-  expect(backgroundPage).toBeTruthy();
-  expect(context.backgroundPages()).toContain(backgroundPage);
-  expect(context.pages()).not.toContain(backgroundPage);
-  await context.close();
-  expect(context.pages().length).toBe(0);
-  expect(context.backgroundPages().length).toBe(0);
-});
-
-it('should return background pages when recording video', async ({ browserType, asset, isHeadlessShell }, testInfo) => {
-  it.skip(isHeadlessShell, 'Headless Shell has no support for extensions');
-
-  const extensionPath = asset('simple-extension');
-  const extensionOptions = {
-    args: [
-      `--disable-extensions-except=${extensionPath}`,
-      `--load-extension=${extensionPath}`,
-    ],
-    recordVideo: {
-      dir: testInfo.outputPath(''),
-    },
-  };
-  const context = await browserType.launchPersistentContext('', extensionOptions);
-  const backgroundPages = context.backgroundPages();
-  const backgroundPage = backgroundPages.length
-    ? backgroundPages[0]
-    : await context.waitForEvent('backgroundpage');
-  expect(backgroundPage).toBeTruthy();
-  expect(context.backgroundPages()).toContain(backgroundPage);
-  expect(context.pages()).not.toContain(backgroundPage);
-  await context.close();
-});
-
-it('should support request/response events when using backgroundPage()', async ({ browserType, asset, server, isHeadlessShell }) => {
-  it.skip(isHeadlessShell, 'Headless Shell has no support for extensions');
-
-  server.setRoute('/empty.html', (req, res) => {
-    res.writeHead(200, { 'Content-Type': 'text/html', 'x-response-foobar': 'BarFoo' });
-    res.end(`<span>hello world!</span>`);
-  });
-  const extensionPath = asset('simple-extension');
-  const extensionOptions = {
-    args: [
-      `--disable-extensions-except=${extensionPath}`,
-      `--load-extension=${extensionPath}`,
-    ],
-  };
-  const context = await browserType.launchPersistentContext('', extensionOptions);
-  const backgroundPages = context.backgroundPages();
-  const backgroundPage = backgroundPages.length
-    ? backgroundPages[0]
-    : await context.waitForEvent('backgroundpage');
-  await backgroundPage.waitForURL(/chrome-extension\:\/\/.*/);
-  const [request, response, contextRequest, contextResponse] = await Promise.all([
-    backgroundPage.waitForEvent('request'),
-    backgroundPage.waitForEvent('response'),
-    context.waitForEvent('request'),
-    context.waitForEvent('response'),
-    backgroundPage.evaluate(url => fetch(url, {
-      method: 'POST',
-      body: 'foobar',
-      headers: { 'X-FOOBAR': 'KEKBAR' }
-    }), server.EMPTY_PAGE),
-  ]);
-  expect(request).toBe(contextRequest);
-  expect(response).toBe(contextResponse);
-  expect(request.url()).toBe(server.EMPTY_PAGE);
-  expect(request.method()).toBe('POST');
-  expect(await request.allHeaders()).toEqual(expect.objectContaining({ 'x-foobar': 'KEKBAR' }));
-  expect(request.postData()).toBe('foobar');
-
-  expect(response.status()).toBe(200);
-  expect(response.url()).toBe(server.EMPTY_PAGE);
-  expect(response.request()).toBe(request);
-  expect(await response.text()).toBe('<span>hello world!</span>');
-  expect(await response.allHeaders()).toEqual(expect.objectContaining({ 'x-response-foobar': 'BarFoo' }));
-
-  await context.close();
-});
-
-it('should report console messages from content script', {
-  annotation: { type: 'issue', description: 'https://github.com/microsoft/playwright/issues/32762' }
-}, async ({ browserType, asset, server, isHeadlessShell }) => {
-  it.skip(isHeadlessShell, 'Headless Shell has no support for extensions');
-
-  const extensionPath = asset('extension-with-logging');
-  const extensionOptions = {
-    args: [
-      `--disable-extensions-except=${extensionPath}`,
-      `--load-extension=${extensionPath}`,
-    ],
-  };
-  const context = await browserType.launchPersistentContext('', extensionOptions);
-  const page = await context.newPage();
-  const consolePromise = page.waitForEvent('console', e => e.text().includes('Test console log from a third-party execution context'));
-  await page.goto(server.EMPTY_PAGE);
-  const message = await consolePromise;
-  expect(message.text()).toContain('Test console log from a third-party execution context');
-  await context.close();
-});
-
 it('should not create pages automatically', async ({ browserType }) => {
   const browser = await browserType.launch();
   const browserSession = await browser.newBrowserCDPSession();
